requirements.txt
aiosmtpd
python-dotenv
pymongo
Flask
eml_parser